Calculates interest with one time credits / debits ...also accidentally split stuff out into separate files, which was supposed to be in a different commit, but whatever
- TODO: add more verified, true test cases
- TODO: test continuously compounded
